+++ roseau/load_flow_single/models/loads.py
@@ -1,8 +1,9 @@
 import logging
 from abc import ABC
-from typing import Final, TypeVar
+from typing import Final
 
 import numpy as np
+from typing_extensions import TypeVar
 
 from roseau.load_flow import RoseauLoadFlowException, RoseauLoadFlowExceptionCode
 from roseau.load_flow.typing import Complex, Id, JsonDict
@@ -14,10 +15,10 @@ from roseau.load_flow_single.models.flexible_parameters import FlexibleParameter
 
 logger = logging.getLogger(__name__)
 
-_CyL = TypeVar("_CyL", bound=CyLoad)
+_CyL_co = TypeVar("_CyL_co", bound=CyLoad, default=CyLoad, covariant=True)
 
 
-class AbstractLoad(BaseConnectable[_CyL], ABC):
+class AbstractLoad(BaseConnectable[_CyL_co], ABC):
     """An abstract class of an electric load."""
 
     element_type: Final = "load"